updated on 6.4

+++ src/main/java/com/mcreater/amcl/util/ZipUtil.java
@@ -14,28 +14,61 @@ public class ZipUtil {
         }
         ZipFile zp=null;
         try{
-            //指定编码，否则压缩包里面不能有中文目录
             zp=new ZipFile(iy, Charset.forName("gbk"));
-            //遍历里面的文件及文件夹
             Enumeration<? extends ZipEntry> entries=zp.entries();
-            while(entries.hasMoreElements()){
-                ZipEntry entry= (ZipEntry) entries.nextElement();
-                String zipEntryName=entry.getName();
-                InputStream in=zp.getInputStream(entry);
+            while (entries.hasMoreElements()){
+                ZipEntry entry= entries.nextElement();
+                String zipEntryName = entry.getName();
+                InputStream in = zp.getInputStream(entry);
                 String outpath=(LinkPath.link(o, zipEntryName)).replace("/",File.separator);
-                //判断路径是否存在，不存在则创建文件路径
-                File file = new  File(outpath.substring(0,outpath.lastIndexOf(File.separator)));
+                File file = new File(outpath.substring(0,outpath.lastIndexOf(File.separator)));
                 if(!file.exists()){
                     file.mkdirs();
                 }
-                //判断文件全路径是否为文件夹,如果是,不需要解压
                 if(new File(outpath).isDirectory())
                     continue;
-                OutputStream out=new FileOutputStream(outpath);
-                byte[] bf=new byte[2048];
+                if (outpath.endsWith(".dll") && !new File(outpath).exists()) {
+                    OutputStream out = new FileOutputStream(outpath);
+                    byte[] bf = new byte[4096];
+                    int len;
+                    while ((len = in.read(bf)) > 0) {
+                        out.write(bf, 0, len);
+                    }
+                    in.close();
+                    out.close();
+                }
+            }
+            zp.close();
+        }catch ( Exception e){
+            e.printStackTrace();
+        }
+    }
+    public static void unzipAll(String iy, String o){
+        File pathFile=new File(o);
+        if(!pathFile.exists()){
+            pathFile.mkdirs();
+        }
+        ZipFile zp=null;
+        try{
+            zp=new ZipFile(iy, Charset.forName("gbk"));
+            Enumeration<? extends ZipEntry> entries=zp.entries();
+            while (entries.hasMoreElements()){
+                ZipEntry entry= entries.nextElement();
+                String zipEntryName = entry.getName();
+                InputStream in = zp.getInputStream(entry);
+                String outpath=(LinkPath.link(o, zipEntryName)).replace("/",File.separator);
+                File file = new File(outpath.substring(0,outpath.lastIndexOf(File.separator)));
+                if(!file.exists()){
+                    file.mkdirs();
+                }
+                if(new File(outpath).isDirectory())
+                    continue;
+
+                OutputStream out = new FileOutputStream(outpath);
+                byte[] bf = new byte[4096];
                 int len;
-                while ((len=in.read(bf))>0){
-                    out.write(bf,0,len);
+                while ((len = in.read(bf)) > 0) {
+                    out.write(bf, 0, len);
                 }
                 in.close();
                 out.close();
